refactor: replace RouterHooks in hooks directory
